+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Complete Xtensa Rust versions when provided one is incompleted (#363)
+- Create $RUSTUP_HOME/tmp if needed (#365)
 
 ### Removed